refactor(storage): split service facade by capability
(cherry picked from commit 6a0f8178ad10309ab94fdb3d40ad9ba1977a4b0f)
